# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ignis API
  version: 1.0.0
  description: |
    ignis calculates annual heating energy demand (`q_h_nd`, in kWh/(m²·a))
    for European building typologies, following EN ISO 13790 over the TABULA
    building dataset. Given a TABULA variant code, it returns the demand and the
    underlying parameters. All endpoints are read-only.

    ## Authentication

    ignis carries no credential of its own and asks for none. Every endpoint
    below is reachable by anything that can open a connection to it, so a
    deployment decides who that is at the network layer. Deploy it only on a
    network whose access you control, behind a platform that has already
    authenticated the end user.

    If the system calling ignis has its own user-facing login (EnerPlanET,
    for example, uses Keycloak), that login authenticates the user *to that
    system*; ignis never sees it. That system's backend then calls ignis on
    the user's behalf.

    Do not publish ignis on a public interface.
  license:
    name: MIT
    url: https://github.com/thd-spatial-ai/ignis/blob/main/LICENSE
  contact:
    name: THD-Spatial-AI
    url: https://github.com/thd-spatial-ai/

# Explicitly no security scheme. ignis carries no credential and asks for none;
# reachability is decided at the network layer (ADR-004). The empty list states
# that, and is what satisfies redocly's security-defined rule in the docs
# workflow.
security: []
servers:
  - url: http://localhost:8088
...
    variables:
...
paths:
...
  /api/v1/calculate/{code}:
    post:
      tags: [Calculation]
      operationId: calculate
      summary: Run the heating-demand pipeline for a variant
      parameters:
        - $ref: "#/components/parameters/VariantCode"
      requestBody:
        required: false
        description: Optional overrides. Omit the body to use the TABULA defaults.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalculateRequest"
      responses:
        "200":
          description: Calculated annual heating demand
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
components:
  parameters:
...
    VariantCode:
      name: code
      in: path
      required: true
      description: >
        TABULA variant code. The first two characters are the ISO 3166-1 alpha-2
        country code, which selects the database table.
      schema:
        type: string
      example: DE.N.SFH.01.Gen.ReEx.001.001
  responses:
    BadRequest:
      description: >
        Invalid input (unknown country, malformed code, a match request missing
        type or giving neither or both of period and year or a non-integer year,
        or an invalid CalculateRequest override (A_ref, h_room, n_Storey or c_m)
        not greater than 0, HeatingDays, I_Sol_*, n_air_infiltration, or
        n_air_use negative, a thermal-bridging ΔU negative, or a surface with an
        unknown type / non-positive area / non-positive u_value)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: no TABULA dataset configured for XX
    NotFound:
      description: Variant code not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Variant not found
    ServerError:
      description: Pipeline failed or returned a non-finite result
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    CalculateRequest:
      type: object
      description: >
        Every field is an optional override of the corresponding TABULA record
        value; omit a field to keep its TABULA default. Fields map 1:1 onto the
        AdvancedParameters groups documented in /api/v1/fields (ClimateConditions,
        SolarGains, ThermalBridges) plus A_ref, the floor area override.
      properties:
        A_ref:
          type: number
          format: double
          description: >
            Overrides the reference floor area (A_C_Ref_Input) from the TABULA
            record. Must be greater than 0.
          example: 150.0
        HeatingDays:
          type: integer
          description: Overrides ClimateConditions.HeatingDays. Must not be negative.
          example: 200
        Theta_e:
          type: number
          format: double
          description: Overrides ClimateConditions.Theta_e (external design temperature, °C). Unbounded.
          example: -5.0
        theta_i:
          type: number
          format: double
          description: Overrides ClimateConditions.Theta_i (internal design temperature, °C). Unbounded.
          example: 20.0
        I_Sol_South:
          type: number
          format: double
          description: Overrides SolarGains.I_Sol_South (Wh/m²·a). Must not be negative.
          example: 400.0
        I_Sol_East:
          type: number
          format: double
          description: Overrides SolarGains.I_Sol_East (Wh/m²·a). Must not be negative.
          example: 150.0
        I_Sol_West:
          type: number
          format: double
          description: Overrides SolarGains.I_Sol_West (Wh/m²·a). Must not be negative.
          example: 150.0
        I_Sol_North:
          type: number
          format: double
          description: Overrides SolarGains.I_Sol_North (Wh/m²·a). Must not be negative.
          example: 80.0
        I_Sol_Hor:
          type: number
          format: double
          description: Overrides SolarGains.I_Sol_Horizontal (Wh/m²·a). Must not be negative.
          example: 500.0
        delta_U_ThermalBridging_Original:
          type: number
          format: double
          description: >
            Overrides ThermalBridges.Delta_U_ThermalBridging_Original (W/m²K,
            added to the envelope's U-values). Must not be negative.
          example: 0.1
        delta_U_ThermalBridging_Refurbished:
          type: number
          format: double
          description: >
            Overrides ThermalBridges.Delta_U_ThermalBridging_Refurbished (W/m²K,
            added to the envelope's U-values). Must not be negative.
          example: 0.05
        h_room:
          type: number
          format: double
          description: >
            Overrides BuildingAppearance.H_room (m), the archetype's assumed room
            height. Feeds the ventilation heat transfer coefficient directly,
            often a large lever, since TABULA's generic default is frequently a
            poor match for a specific real building. Must be greater than 0.
          example: 2.8
        n_Storey:
          type: integer
          description: >
            Overrides BuildingAppearance.N_Storey, the archetype's assumed storey
            count. Accepted and validated, but without effect on any shipped
            TABULA variant: it reaches the result only through the envelope-area
            estimation path, which applies when Code_TypeIntake_EnvelopeArea is
            "Estimation", and no shipped variant carries that code. Its other
            consumers are the envelope plausibility ratios, which end in
            Check_EnvArea_ExactToEstim, and nothing reads that.
            Must be greater than 0.
          example: 3
        n_air_infiltration:
          type: number
          format: double
          description: >
            Overrides AirInfiltration.N_air_infiltration (1/h), the uncontrolled
            air-change rate through gaps and cracks. Must not be negative.
          example: 0.2
        n_air_use:
          type: number
          format: double
          description: >
            Overrides AirInfiltration.N_air_use (1/h), the ventilation air-change
            rate during occupancy. Must not be negative.
          example: 0.4
        c_m:
          type: number
          format: double
          description: >
            Overrides HeatTransfer.C_m (J/(m²K)), the internal heat capacity per m²
            of useful floor area. Feeds the gain-utilization time constant
            (calc_level_12.go). Must be greater than 0.
          example: 165000
        surfaces:
          type: array
          description: >
            Replaces TABULA's fixed 2-3 slots per element category with an
            arbitrary list of individual physical surfaces, as a real
            building's geometry has rather than a generic archetype.
            Surfaces are grouped by type and collapsed into an area-weighted
            equivalent (summed area, area-weighted average U-value) before
            the pipeline runs. This is mathematically exact, since U-values are
            conductances and conductances in parallel add weighted by area.
            A category with no surfaces given keeps its TABULA default.
            The caller owns the geometric accuracy of the areas, U-values
            and orientations it sends; ignis does not cross-check them
            against the archetype.
          items:
            $ref: "#/components/schemas/Surface"
    Surface:
      type: object
      description: >
        One physical building element. Shape mirrors BuEM's
        envelope.elements[] so the same source data can drive both models.
      properties:
        id:
          type: string
          description: Caller-assigned identifier, used only in error messages.
          example: wall-1
        type:
          type: string
          enum: [roof, wall, floor, window, door]
        area:
          type: number
          format: double
          description: m². Must be greater than 0.
          example: 60.0
        u_value:
          type: number
          format: double
          description: W/(m²K). Must be greater than 0.
          example: 0.8
        azimuth:
          type: number
          format: double
          description: >
            Degrees, 0=N/90=E/180=S/270=W. Windows only: rounds to the
            nearest of ignis's five solar-gain buckets (N/E/S/W/Horizontal).
            Defaults to South (180) if omitted.
          example: 180
        tilt:
          type: number
          format: double
          description: >
            Degrees from horizontal, windows only: 0 is a flat skylight, 90 a
            vertical window. Within 5° of 0 the window is bucketed as
            Horizontal regardless of azimuth. This is the inverse of
            city2tabula's surface tilt (0 = vertical wall, 90 = flat roof); a
            caller passing city2tabula geometry must convert first with
            ignis_tilt = 90 - c2t_tilt.
          example: 90
      required: [id, type, area, u_value]
    CalculateResponse:
      type: object
      properties:
        variant_code:
          type: string
          example: DE.N.SFH.01.Gen.ReEx.001.001
        q_h_nd:
          type: number
          format: double
          description: Annual net heating energy demand
          example: 282.72
        unit:
          type: string
          example: kWh/(m2.a)
      required: [variant_code, q_h_nd, unit]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]